refactor: dispatching functions
