+++ src/scripts/units/interface/uiEnhancer/phind/changeExisting/improveStyles.js
@@ -20,3 +20,13 @@ function modifyingStyle() {
     elt.classList.remove("mb-7");
   });
 }
+
+function addStyle() {
+  let styleTag = document.createElement('style');
+  document.querySelector("body").appendChild(styleTag);
+  let stylesheet = styleTag.sheet;
+
+  stylesheet.insertRule(".smallScreens { display: none!important; }", 0);
+  stylesheet.insertRule("@media screen and (max-width: 1025px) { .smallScreens { display: inline-block!important; } }", 1);
+}
+
